Plot movement (#149)
* Rough Draft Figure Alignment

* Resize Figure

* Move Everything to Figure 2

* Update figure2.py

* Update figure2.py

* Update figure2.py
